Largest prime number

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	unsigned int long n = 612852475143, a = (int) sqrt(n);
+int n = 612852475143, a = (int) sqrt(n);
 
 while (1)
 {
